chore(QOV-1876): extend lint scope (#566)
* chore: update GoReleaser configuration and dependencies

* chore(QOV-1875): update dependencies and remove gosimple linter from configuration

* chore: update Go version to 1.26.2

* chore(QOV-1876): apply golangci-lint --fix to qovery and client packages

* chore(QOV-1876): extend lint scope to qovery and client with legacy excludes

* chore(QOV-1876): enable unconvert linter on qovery and client

* chore(QOV-1876): enable gocritic linter on qovery and client

* chore(QOV-1876): enable unparam linter on qovery and client

* chore(QOV-1876): enable unused linter on qovery and client

+++ qovery/data_source_job.go
@@ -3,6 +3,7 @@ package qovery
 import (
 	"context"
 	"fmt"
+
 	"github.com/qovery/qovery-client-go"
 
 	"github.com/AlekSi/pointer"
@@ -105,13 +106,13 @@ func (d jobDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, res
 			"max_duration_seconds": schema.Int64Attribute{
 				Description: descriptions.NewInt64MinDescription(
 					"Job's max duration in seconds.",
-					int64(job.MinDurationSeconds),
-					pointer.ToInt64(int64(job.DefaultMaxDurationSeconds)),
+					job.MinDurationSeconds,
+					pointer.ToInt64(job.DefaultMaxDurationSeconds),
 				),
 				MarkdownDescription: descriptions.NewInt64MinDescription(
 					"Job's max duration in seconds.",
-					int64(job.MinDurationSeconds),
-					pointer.ToInt64(int64(job.DefaultMaxDurationSeconds)),
+					job.MinDurationSeconds,
+					pointer.ToInt64(job.DefaultMaxDurationSeconds),
 				),
 				Optional: true,
 				Computed: true,
@@ -119,13 +120,13 @@ func (d jobDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, res
 			"max_nb_restart": schema.Int64Attribute{
 				Description: descriptions.NewInt64MinDescription(
 					"Job's max number of restarts.",
-					int64(job.MinNbRestart),
-					pointer.ToInt64(int64(job.DefaultMaxNbRestart)),
+					job.MinNbRestart,
+					pointer.ToInt64(job.DefaultMaxNbRestart),
 				),
 				MarkdownDescription: descriptions.NewInt64MinDescription(
 					"Job's max number of restarts.",
-					int64(job.MinNbRestart),
-					pointer.ToInt64(int64(job.DefaultMaxNbRestart)),
+					job.MinNbRestart,
+					pointer.ToInt64(job.DefaultMaxNbRestart),
 				),
 				Optional: true,
 				Computed: true,
